requirements.txt
alembic==1.4.2
apispec==4.0.0
apispec-webframeworks==0.5.2
astroid==2.4.2
attrs==20.2.0
Babel==2.8.0
boto3==1.16.25
botocore==1.19.25
certifi==2020.6.20
cffi
chardet==3.0.4
click==7.1.2
coverage==6.4
cryptography
ecdsa==0.14.1
et-xmlfile==1.0.1
Flask==1.1.2
flask-apispec==0.10.0
Flask-Babel==2.0.0
Flask-Cors==1.10.3
Flask-Inputs==0.3.0
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
future==0.18.2
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
isort==5.4.2
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jsonschema==3.0.1
lazy-object-proxy==1.4.3
langdetect==1.0.8
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
more-itertools==8.5.0
openpyxl==3.0.5
packaging==20.4
pluggy==0.13.1
py==1.9.0
py-linq==1.2.4
pyasn1==0.4.8
pycparser==2.20
PyMySQL==0.10.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==7.1.2 
pytest-cov==3.0.0
pytest-mock==3.3.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-jose==3.2.0
python-keycloak==0.22.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rsa==4.6
sendgrid==6.9.7
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
webargs==6.1.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.3.1
marshmallow_enum==1.5.1
PyJWT==2.0.1
xmltodict==0.12.0
expiringdict==1.2.1
# flask_httpauthc
# freezeguns
twilio
# jwt
# PyJWT==2.1.0
flask-unittest==0.1.3
# coverage